+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Use Semantic Highlighting instead of decoration styling for conditions and actions that affect quest workflow, action and symbol parameters and symbols that don't follow standard syntax and can't benefit of syntax highlighting.
 
+### Fixed
+
+- Fixed action `change foe`.
+
 ## 0.14.0
 
 ### Improved